Settings: Advanced low battery indicator options (1/2)
Based on a idea I saw on Liquid and from CR3C0 (thanks guys) which we
modified to an extended version
This commit introduces following new SlimBean modes to indicate low battery
     * 0 = show dialog + play sound (default)
     * 1 = fire notification + play sound
     * 2 = show dialog only
     * 3 = fire notification only
     * 4 = play sound only
     * 5 = none

Due that notification is a unique SlimBean feature some words on it. Notification
behaviour is the same as Pop-up. User can dismiss it manually or it is
dismissed automatically if charger is connected, click on notification opens
PowerUsageSummary. Also you get a status bar ticker message about the notification
and notification is always on first place of list. And last if vibration is enabled
notification vibrates.

Change-Id: I372c566a2d84b04856d0be518a894c3699d00f8a

Conflicts:
	res/values/gummy_arrays.xml
	res/values/gummy_strings.xml